refactor: remove approval pass-through modules
