+++ packages/cli/src/commands/_helpers/runInstructions.ts
@@ -4,7 +4,7 @@
  * notices, and the tool-use/multi-agent instruction assemblers built on top
  * of them. Formerly one file per concern under `_helpers/`.
  */
-import type { ApprovalInstructionContext } from '@cli/runtime/approvalPolicyAvailability';
+import type { ApprovalInstructionContext } from '@cli/runtime/approval/approvalPolicy';
 
 const TERMINAL_RUN_GUIDANCE =
   'This CLI run exits after your final response. Do not end by asking the user whether to perform more work; either complete the requested work now or state the exact artifacts and next command/action for a future run.';
