+++ packages/cli/src/runtime/runExecution.ts
@@ -17,7 +17,7 @@ import { EXECUTION_STATUS, RUN_OUTCOME } from '@shared/schemas';
 import { generateExecutionId } from '@utils/core';
 import { toErrorMessage } from '@utils/errors/errorMessage';
 
-import { approvalPromptsUnavailable } from './approvalPolicyAvailability';
+import { approvalPromptsUnavailable } from './approval/approvalPolicy';
 import { createHeadlessCliHostInteractions } from './approvalAdapter';
 import { finalizeCliExecution } from './executionFinalization';
 import { attachCliSessionProgressProjection } from './sessionProgressSubscription';
